+++ src/constants.ts
@@ -145,6 +145,7 @@ export const PACKAGES_TO_GET_INFO_FOR = [
   'lodash-es',
   'qunit',
   ...CHECKED_LODASH_METHODS.map((method) => `lodash.${method}` as const),
+  'rxjs',
 
   'astro',
   'vue',